+++ src/main/java/org/bytedeco/javacpp/CLongPointer.java
@@ -44,6 +44,9 @@ public class CLongPointer extends Pointer {
     public CLongPointer(long size) {
         try {
             allocateArray(size);
+            if (size > 0 && address == 0) {
+                throw new OutOfMemoryError("Cannot allocate new CLongPointer(" + size + ")");
+            }
         } catch (UnsatisfiedLinkError e) {
             throw new RuntimeException("No native JavaCPP library in memory. (Has Loader.load() been called?)", e);
         }