+++ src/main/java/org/bytedeco/javacpp/LongPointer.java
@@ -66,6 +66,9 @@ public class LongPointer extends Pointer {
     public LongPointer(long size) {
         try {
             allocateArray(size);
+            if (size > 0 && address == 0) {
+                throw new OutOfMemoryError("Cannot allocate new LongPointer(" + size + ")");
+            }
         } catch (UnsatisfiedLinkError e) {
             throw new RuntimeException("No native JavaCPP library in memory. (Has Loader.load() been called?)", e);
         }