+++ src/main/java/org/bytedeco/javacpp/PointerPointer.java
@@ -116,6 +116,9 @@ public class PointerPointer<P extends Pointer> extends Pointer {
     public PointerPointer(long size) {
         try {
             allocateArray(size);
+            if (size > 0 && address == 0) {
+                throw new OutOfMemoryError("Cannot allocate new PointerPointer(" + size + ")");
+            }
         } catch (UnsatisfiedLinkError e) {
             throw new RuntimeException("No native JavaCPP library in memory. (Has Loader.load() been called?)", e);
         }