+++ src/main/java/org/bytedeco/javacpp/SizeTPointer.java
@@ -43,6 +43,9 @@ public class SizeTPointer extends Pointer {
     public SizeTPointer(long size) {
         try {
             allocateArray(size);
+            if (size > 0 && address == 0) {
+                throw new OutOfMemoryError("Cannot allocate new SizeTPointer(" + size + ")");
+            }
         } catch (UnsatisfiedLinkError e) {
             throw new RuntimeException("No native JavaCPP library in memory. (Has Loader.load() been called?)", e);
         }